Workaround for differences between touch screen and mouse.
There are concurrency issues when defining events.

e.g.

If I tapping a node selects it.
If I use a touch screen it thinks I tapped a node that is already
selected.
If I instead use a mouse it correctly doesn't think the note is
selected, unless I tap it again.