Style + play connections
The style of a page, in the form of inner html is stored in the
cytoscape node.

This same inner html is used to generate the page in the play module.

Decisions (grabbed from edges) are now dynamically created buttons in
the page view.

Hide/show edit pane functions moved to pageOverlay.js, since the overlay
is now where editting is done.