Node number fix, Edges now labelled, Handles now have dynamic names.
Handles will pull from name of container object in cytoscape each time
the overlay is loaded, rather than being "set" when a container is
created. This is useful as it means we can now delete and update names
of containers easily without having to regex through the html attribute
and change values.