Z-index and bug fixes.
Z-index "Bring to Front"/"Send to Back" option added for default
containers.
Playpage bug fix (the presence of containers in it was messing with
.length attribute retrievals, so now I wipe it's contents when we leave
the tab).
Resolution bug fix (calculating the center of the page overlay was
messing up when the page overlay was not open).
Generalising some of the context menu entries, so that it's easier to
use the same parts e.g. bring to front for both img containers and text
containers