Compound collapse/expand/add. Page overlay scale.
- I’ve been involved with the interact.js community for the duration of
this project, and someone brought up and issue that I answered, and
which prompted some ideas of my own. (See
https://github.com/taye/interact.js/issues/430 for our correspondence)
o Using the transform:scale css property I was able to devise a solution
to designing pages with changing resolutions, and translating these
resolutions changes (while maintaining appropriate positions) from the
creator to the player.
o Essential from the creator view the window will maintain its aspect
ratio, and try to fill up as much space as possible. In the play module
it will do the same, so that it will expand to be as large as possible
to fit the user’s screen.
o If a creator wants to develop for a different aspect ratio (e.g.
mobile) they can specify that in project settings.

o There were some issues with blurry icons when scaling that may need to
be investigated.
- Fixed a bug where the play module would indefinitely wait to load if
no audio assets were defined.
- Added functionality for clicking the “add node” etc… buttons to revert
to default state when clicked again. Slightly trickier for the dropdown
buttons.
- Utilized the mymodal.prompt I developed in audio adding.
- Extended hotkey support to mymodal (enter = confirm, escape =
cancel+exit)
- On Darryl’s feedback I developed functionality for compound nodes to
be collapsed and expanded by clicking on their label.
o Attempted to use an existing extension in cytoscape, but it would
throw errors with our implementation.
o Encountered an issue in my own functionality: I wanted nodes to be
hidden when a node was collapsed, but when this was done the collapsed
node could not be moved. The solution worked if the opacity of a
compound node’s children was set to 0.0001 (practically transparent),
however those children could still technically be moved, and thus would
change the size of the collapsed parent node.
o This turned out to be a bug in cytoscape that has been fixed since we
last updated the library, so after investigation I updated our version
of cytoscape.
o Afterwards I ensured compliance with other features (e.g. an edge
can’t have a compound node as a source or target, but it can have a
compound node’s child as once).

Expanded

Collapsed
- Nodes (and even compound nodes) can be dragged into compound nodes to
add them as a child to the compound node.
- Note that recursively collapsing compound nodes isn’t retaining the
children’s relative position correctly due to the changes in the size of
a parent when a child expands or collapses.